+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 unreleased
 ==========
+* feat: Compliance number filter added
 
 1.0.0 (2022-04-25)
 ==================